refactor(autocode-auth): migrate middleware.ts to Next.js 16 proxy convention
Rename middleware.ts → proxy.ts per Next.js 16's deprecation of the
middleware file convention. The proxy API is identical — same function
signature, same config.matcher, same NextAuth auth() wrapper. This is
a pure file rename with no logic changes.

Eliminates the build warning:
"The middleware file convention is deprecated. Please use proxy instead."

103 tests passing, build clean with zero warnings.
